+++ Vec2.h
@@ -12,7 +12,7 @@ public:
 		:x(0), y(0)
 	{}
 
-	Vec2(double x, double y)
+	Vec2(float x, float y)
 		:x(x), y(y)
 	{}
